requirements.txt
fastapi
uvicorn
gunicorn
sqlalchemy
psycopg2-binary
bcrypt
python-jose[cryptography]
pydantic[email]
pydantic-settings
python-dotenv
alembic